fix: code quality : refactoring into subdirectories
