--- constants/Algorithms.py
@@ -1,4 +1,0 @@
-class Algorithms:
-    NAIVE_ALGORITHM = 'Naive algorithm'
-    ANT_COLONY_ALGORITHM = 'Ant colony algorithm'
-    EVOLUTIONAL_ALGORITHM = 'Evolutional algorithm'
