--- models/Algorithm.py
@@ -1,4 +1,0 @@
-class Algorithm:
-    def __init__(self) -> None:
-        parameters = None
-        name = None
