+++ views/ScrollingListView.py
@@ -5,7 +5,10 @@ from constants.Colors import Colors
 
 from models.algorithms.Algorithm import Algorithm
 
-from models.Algorithm import Algorithm
+from constants.Config import GRAPH_WINDOW_WIDTH
+from constants.Colors import Colors
+
+from models.algorithms.Algorithm import Algorithm
 
 class ScrollingListView:
     def __init__(self, screen, x, y, width, height) -> None:
